requirements.txt
# ╔══════════════════════════════════════════════════════════╗
# ║              RONIN-V — Python Dependencies              ║
# ╚══════════════════════════════════════════════════════════╝

# ─── Core LLM ───
ollama>=0.4.0

# ─── Memory ───
chromadb>=0.5.0

# ─── Terminal UI ───
rich>=13.0
prompt-toolkit>=3.0

# ─── System ───
psutil>=6.0

# ─── Config ───
pyyaml>=6.0
click>=8.0
colorama>=0.4
